refactor: Renamed AppRouter to Main refactor: Change Option to non-exported component chore: Fixed potential issue in Input component chore: Changed Row to div rather than span
